+++ adhara/lib/datainterface/models/decorators.dart
@@ -1,0 +1,16 @@
+
+class AdharaModelField {
+//  /// Whether the field is serialized. Defaults to `true`.
+//  ///
+//  /// If a field is not serialized, it must either be `@nullable` or specify a
+//  /// default for deserialization to succeed.
+//  final bool serialize;
+//
+//  /// The wire name when the field is serialized. Defaults to `null` which
+//  /// indicates the name is to be taken from the literal field name.
+//  final String wireName;
+
+  const AdharaModelField(
+//      {this.compare = true, this.serialize = true, this.wireName}
+  );
+}